share icon tests complete
