+++ frontend/src/jsconfig.json
@@ -1,0 +1,3 @@
+{
+    "typeAcquisition": {"include": ["jest"]}
+}
