+++ package.json
@@ -20,7 +20,10 @@
         "transform": {
             "^.+\\.jsx?$": "babel-jest",
             "^.+\\.svg$": "jest-svg-transformer"
-        }
+        },
+        "modulePathIgnorePatterns": [
+            "<rootDir>/backend/"
+        ]
     },
     "keywords": [],
     "author": "",
@@ -48,7 +51,9 @@
         "@babel/runtime": "^7.12.5",
         "@testing-library/jest-dom": "^5.11.9",
         "@testing-library/react": "^11.2.3",
+        "@testing-library/user-event": "^12.6.2",
         "aws-sdk": "^2.828.0",
+        "babel-eslint": "^10.1.0",
         "babel-jest": "^26.6.3",
         "babel-loader": "^8.2.2",
         "cors": "^2.8.5",
